Remocao do pysolr de alguns requirements e criacao de um modulo de log pro solr
