+++ extract_hierarchy/requirements.txt
@@ -1,7 +1,2 @@
 hdfs
 argparse
-impyla==0.15.0
-thrift==0.9.3
-pysolr
-requests-kerberos
-kerberos
