+++ ftp_detran/requirements.txt
@@ -1,6 +1,1 @@
 argparse
-impyla==0.15.0
-thrift==0.9.3
-pysolr
-requests-kerberos
-kerberos
