+++ ftp_tce/requirements.txt
@@ -7,4 +7,3 @@ unidecode
 pysolr
 requests-kerberos
 kerberos
-
